requirements.txt
# For sending notifications
python-telegram-bot

# For loading .env files
python-dotenv==1.0.1

# For data manipulation
pandas

# For making HTTP requests
requests==2.31.0

# For parsing HTML
beautifulsoup4

# For data models (Pydantic v2)
pydantic>=2.6.0,<3.0.0

# For LLM extraction (OpenAI SDK v1)
openai>=1.0.0

# For database access and migrations
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
